requirements.txt
accelerate==1.10.1
datasets==4.1.1
evaluate==0.4.6
numpy==2.3.3
peft==0.17.1
PyYAML==6.0.2
PyYAML==6.0.3
torch==2.6.0+cu124
tqdm==4.67.1
transformers==4.55.4
